feat: Make FEED_URL configurable
This commit makes the FEED_URL configurable via the .env file.

Previously, the FEED_URL was hardcoded in the index.js file. This made it difficult to change the feed URL without modifying the code.

With this change, the FEED_URL can now be set in the .env file. This makes it easier to change the feed URL without having to modify the code.

A few example FEED_URLs have been added to the .env.example file for reference.